Replace nn.Embedding to nn.Linear

+++ test/transform/test_toy.py
@@ -16,7 +16,7 @@ def test_collate():
     collate = Collate(dataset.tokens, 2, 10, 2)
     for x in dataset:
         t0, t1, t2, t3, out, bsize = collate([x])
-        assert t0.dtype == torch.int64
+        assert t0.dtype == torch.float32
         assert t1.dtype == torch.bool
         assert t2.dtype == torch.float
         assert t3.dtype == torch.bool
